Added mail-notify gem
Basic setup for mail-notify gem with a dummy example of how we can use a
template in the gov notify dashboard and inject our content from the app
into the template with erb format.

This will require a template id which is included in the change but also
an gov notify api key in your .env file which you get it from the
dashboard.