Add documentation for our data model
I've taken a first stab at adding documentation for our intended data
model. At this point it time, the services encapsulated by this
application are in their formative stages, so this data model will
undoubtedly change over time.

Right now I consider this to be a 'draft' document, which shows the
intended data model for the application.