Import GIAS data into GiasSchool table
We need to import GIAS data into our app. The GIAS service does not have
an official API. What they do have is a CSV that they publish every day
and we can download this CSV and parse the data.

To do this, we have introduced a rake task `rake gias_update` which will
download the CSV file for that day as a temp file and upsert the data
into our DB. All in all we have about 24k GiasSchool records created
containing schools and MATS that are still open and in England