Remove deprecation notice from tests
The following deprecation notice was being shown when running the RSpec
test suite:

```
Rails 7.1 has deprecated the singular fixture_path in favour of an array
```

We should set `config.fixture_paths = []` instead of
`config.fixture_path`.

However, we don't actually use ActiveRecord fixtures. We prefer to use
FactoryBot. So I've fixed the deprecation, but then have commented out
the line since it's not something we use.

If we need to enable it in the future, I've fixed the deprecation notice
so we should be able to simply uncomment the line.