Use Rails.logger instead of Logger
Previously with Logger, there was noise added to the terminal when running
the specs. Because of this we want to use Rails.logger instead, in production
and development

Temporarily silent the Active Record logger when
upserting all the GiasSchools. This only adds noise
to the logs when running the gias_update task in production.