Add an API-first Provider model
This integrates with the Provider API in an API-first way – e.g. by
calling out to the API rather than relying on an overnight "ETL" sync
job to cache Provider records in a local Active Record model.

However, I'm putting this work on hold for now because it's become
increasingly clear that an overnight ETL would grant us more flexibility
in how we interact with Provider data. It currently seems to be the more
pragmatic approach, so I'm putting this API-first implementation on ice
for now.