Add a 'placements' flag to Providers
This commit adds a 'placements' boolean (true|false) attribute to the
Provider model. This follows the existing pattern established by the
School model which uses the boolean values to indicate whether that
organisation has been 'onboarded' into the service.

We will use this to differentiate between 'known' providers – i.e.
providers whose name and details we know – and those which have been
onboarded into the School Placements service.

The 'known' providers will be populated using data sourced from the
Teacher Training Courses API, meaning we'll be able to use it as a
lookup table whenever we need to list *all* providers, regardless of
their onboarding state.