Add Faker gem
This will be useful for when we need to generate pseudo-random values
for fields. Useful for when we don't really care what the values of the
fields are, but we want them to be populated with _something_.

For example, this could be used in tests where we need to generate
schools with an assortment of different names.

However, be warned: just because Faker randomly provides values, there
is no guarantee that values will always be unique. If you need
guaranteed unique values, use a 'sequence' with your FactoryBot factory.