Redirect users to organisations index/show after sign in
In claims we need to redirect our users to the organisations index page.
Some users do have only one organisation(school) so for those users we
will redirect them to the school show page, we do this via another
redirect in the schools controller just to make sure a user with one
school doesn't land on the organisation(school) index page.

This change also affects placements where we will always redirect the
user after sign in to the organisations index page.