Allow workflow to push docker images
I've granting 'write' permission for 'packages'. This will allow the
"Build and Deploy" workflow to build and push docker images up to the
container image repository.

Without this, the workflow fails when run on PRs created by Dependabot.
This is because Dependabot itself doesn't have write permission to the
docker repository, causing jobs to fail with the error message:

```
denied: installation not allowed to Write organization package
```

For more context, see [this Slack thread][1].

[1]: https://ukgovernmentdfe.slack.com/archives/C06962474AC/p1705407557674889