Use new onboarding status attributes
This commit updates the various areas of the code which used the
"claims" and "placements" boolean fields, and updated them to use the
new "claims_service" and "placements_service" names.

In addition to that, I've removed the `#claims` and `#placements` scopes
from the top-level `School` model. I've done this to discourage use of
the top-level `School` model – instead, we can use the namespaced
`Claims::School` and `Placements::School` models which are already
scoped to each service.