Rename "placements" attribute for Providers
I've renamed the Provider attribute "placements" to
"placements_service", and have updated associated code accordingly.

As with Schools, the "placements_service" attribute on Providers
indicates that a Provider has been onboarded into the School Placements
service.

This change has been made to free up the `Provider#placements` method,
which is likely to be more useful as an association once we have
modelled Placements within the service.

Note: the Provider model does not have an associated "claims_service"
attribute, because they are never onboarded into that service.