Schools in claims need to be Claims::School
In the claim/belongs_to_school, if the current_user is not a support
user. We will return all the schools for that user. This is technically
correct but the schools that will be returned are of type `School` not
`Claims::School`. This doesn't allow us to access the methods on
`Claims::School` so the easiest fix for now is to use `#becomes` method.
A later refactor might clean this up