Refactor autocomplete JavaScript into stimulus
A refactor has been done to the provider form and school form that
allows the user to add a school or a provider.

These were using 'urn' and 'code' to search a school or a provider in DB, this has
been changed to 'id', to comply with the stimulus autocomplete.

I have also changed the form input name of the school and provider from
'urn' and 'code' to just 'id'