Add DFE sign in integration for login
We need to allow our users to login into our service using DFE sign in.

This commit adds the configuration tu accomodate this, we find the user
by the dfe_sign_in_uid. This should work for placements and claims
through env variables. And for different env, production, staging, local
etc.

We need to create users with dfe sign in emails so that they can login
in our app using dfe sign in. This can be done by a support persona now.

There is no system test for the dfe sign in path, this is because we
need to refactor most of our tests as they use persona sign in, they
cannot work together because the omniauth config loads only once