Refactor system specs to use mock dfe sign in
Now that we have dfe sign in DfE sign in, we should have integration
tests for it. The problem is that the omniauth.rb file that configures
the persona or the DFE sign in path gets loaded only once so we cannot
switch between the two when testing.

We can technically reload the file but rails will complain about trying
to modify frozen strings.

Persona login path is only used for developers, and QA. In production we
will always use DFE sign in so our specs should reflect this, when we test
the login path, we should only test DFE sign in not persona.

This commit adds a dfe_sign_in_user_helper to allow us to easily create
the omniauth payload that dfe sign in needs by passing the user we want
to sign in