Add Rspec smoke tests
Smoke tests should not use local postgres connection

We have a few smoke tests that we run and point them to our qa apps.
Even when changing the Capybara host, rails still wants to connect to a
postgres server locally.

This commit adds nulldb gem which returns a null object instead of
raising an error if rails doesn't find a connection. We use this gem
when the env variable `SMOKE_TEST` is true, so when running the smoke
tests this variable needs to be true and the smoke tests work as
expected.

We're using the nulldb gem because our setup requires us to mock a db
connection running locally even though we point our capybara instance
to a remote server to run these smoke tests.