Temporarily disable some RuboCop rules
I'm temporarily disabling some of the new RuboCop rules for RSpec so we
can incrementally fix-up linting errors in the codebase.

I'm doing this to avoid creating a mega PR that fixes everything all in
one go. This will likely be a nightmare because the codebase will be a
moving target, and it'll be miserable to work on & review.

Plus by taking this approach, it'll be possible to parallelise the
remaining cops so they can be picked up by different developers in the
team.