Stop passing `current_service` into the session
There was some indirection whereby the `current_service` was passed into
the DfESignInUser model via the user's session.

This wasn't actually necessary because we always have access to the
`request.host` whenever we have access to the `request.session` – so we
can always work out the current service based on that.

Not only does it avoid redundant data being stored in the user's session
(which has limited storage capacity).

It also resolves an inconsistency where the value of
`session["service"]` was sometimes a String (if it'd been read from the
session cookie) and sometimes a Symbol (if it'd just been set by
`current_service`). This was causing problems because `DfESignInUser`
expects the service to be a Symbol (for example, see
`DfESignInUser#user_klass`).