Update Azure documentation for new auth
The way we access Azure has changed – we need to grant ourselves access
to 'groups' instead of 'Azure resources' now.

I've updated the documentation and screenshots to reflect this.

I've also removed the note about ignoring errors about `kubelogin` –
this is now required, so it's no longer safe to ignore error messages
about this command missing.