Refactor Claims form
This refactor changes how the claims form works. Now when clicking the
Add claim button, the user will be redirected to the claims#new method
where he needs to choose a provider, this users the
Claims::ProviderForm.

After clicking Continue he will be redirected to claims/mentors#new,
this uses Claims::MentorsForm. This page is where the users selects what
mentors he wants for this claim.

After clicking Continue the user will be redirect to the claims#check
page. From here he can change the provider on the claim or the mentors
on the claim

Both these controller actions also have a create, edit and update
method, these are used when changing the provider or mentors on the check
page, they also redirect the user back to the check page rather then
forcing the user to go through the form again.