Add mentor training hours to claim form
The mentor training hours form will come after the user selects the
mentors for the claim. It will cycle through the mentors that don't have
mentor training hours.

A column has been added to the mentor_training table,
'custom_hours_completed'. This is for capturing the custom hours the
user can input.

After inputting all the hours for the mentors the form will redirect to
the check claim page.

The form also handles the paths of the back links and the success
redirects.