Update the routes_helper.rb and routes_helper_spec.rb
https://trello.com/c/psXhmCl9

Removes the loopy logic where we were accessing a `current_service` due the inclusion of `ApplicationHelper` via `ApplicationController`. This made the code hard to test and generally smelled a little. We now call `HostingService` directly removing this dependency.

This change facilitated the need to rewrite the spec file, which is included in this commit.