Refactor claim statuses
Previously, the claim status is determined by a boolean draft column
on the claims table. As requirements have changed this is not supporting
the functionality we need.

We need an 'internal' status because we create a claim in DB when the
user adds a mentor to a claim. Those claims should not be shown to the
user.

The user should see claims that are draft or submitted, the other two
statuses added in this commit.

We could have used the claim values to calculate the status of the claim
but the logic would not scale very well so an enum seemed like the
correct choice here.