Add shared examples for `ServicePattern`
https://trello.com/c/00ckRT1T

Some of the base functionality of this module was untested, so I created a shared example to make sure all services are behaving as expected.

`NotImplementedError` is generally reserved for the operating system level and inherits from `ScriptError` instead of `StandardError` and does not have the expected behaviour when using `rescue`. I have replaced this for `NoMethodError`.