Add the feedback_page_spec.rb
https://trello.com/c/00ckRT1T

This one was a bit of a nuisance as we haven't implemented this page fully in Placements yet, this does set the foundation for us to extend it and brings up our code coverage however. Removing this page was discussed but due to the existing generic setup between Claims and Placements we may have ended up with more code rather than less.