Claim policies
The claim form didn't have any policies, meaning a claim could have been
updated or submitted twice by a user if they would know the url.

This commit fixes these edge cases.

It also will redirect the user to root path if he doesn't have access to
a resource. This is because redirecting back to request.referrer can cause
a redirection loop if the request.referrer is not a path the user has access
to as well.