Use fixtures for Regions test data
We only support a fixed set of region data: Inner London, Outer London, Fringe, and Rest of England.

To ensure we're only referring to one of the preset regions, we can use fixtures.

As these regions have properties, they still deserve to be database models - we can also use that data for analytics off-app.