Fix back buttons on the claims form
Previously, if a user didn't input mentor hours after adding a mentor
from the check page of the claim form and click the back link. A mentor
without hours would be added to the claim.

This commit will remove any mentors without hours when the user lands on
the check page of a claim. Fixing the issue

Also, this commit will redirect back to the check page when trying to
edit mentor training hours and clicking the back link
if the claim is ready to be checked. Previously the user would need to
cycle through all the mentor hours.