Update documentation for connecting to Azure
The steps for authenticating to Azure have been simplified, so I've
updated the documentation accordingly.

We no longer need a temporary grant to connect to the test cluster.

I'm not sure exactly what the requirements are for connecting to the
production cluster, but we'll need to document that once we're up and
running in that environment.