Split the claim forms into 2
Previously the claim form was used for the support user and normal user
journey. These journeys start to diverge a bit, mainly when redirecting
the user so to make our code a bit more manageable we decided to split
the claim forms into 2.

This commit also fixes some issues with redirecting the support user to
the correct page