Update the search_form partial to be more generic
Use `name` for the search field name, as standard form field helpers use.

Use `value` for the search field input value, as standard form field helpers use. Also default to `params[name]`.

The `label` argument now accepts a hash with `text` and `size`. This is to allow the customisation of the label font size.