Download GIAS CSV files with the correct encoding
GIAS CSV files are published with ISO-8859-1 character encoding.

However the GIAS web server doesn't specify this in the response
headers, meaning we need to set it manually on the downloaded CSV file.

Without this, strange characters appear in imported school data.

This was originally handled in the 'importer' service, but I think it's
actually more appropriate to set the encoding correctly in the
'downloader' service. That way we can keep this encoding quirk 'at the
front door' and avoid any other part of our app needing to know the
specifics of how that file is encoded.