Add a high-level test for Gias::SyncAllSchoolsJob
I've added a high-level integration test for this job because I intend
to split & refactor the underlying service objects, and I want to make
sure I don't break the overall GIAS import feature.

This tests the end to end happy path, starting with the GIAS CSV file
being downloaded, and ending with schools existing in the database.