Support claims index page filters
We need to allow our support users to filter the claims.

Two filters, School and Provider filters have 2 search fields where
you can search by school name or provider to get the filter you want.
This is done by javascript in a stimulus controller.

The other 2 filters, Submitted after and Submitted before filters are
date filters and use the govuk_date_field form helper. You cannot set
the value of this helper as it's sending 3 params not just 1. So to
persist the value of these filters we need a form object.

Claims::Support::Claims::FilterForm is a form object that is meant to
just house the params and all the logic that these filter need.

All the params of the filters are in the url.