Refactor the GIAS importer to follow ETL pattern
Currently, school data is imported from GIAS in two steps:

1. Download the CSV
2. Import the CSV (filtering out unwanted schools in the process)

We need to perform some 'massaging' of the data. For example, the
importer excludes schools which aren't needed within our app. These are
filtered out because either the school has closed, or it isn't in
England (which is where the ITT policy currently applies).

We also have an upcoming need to convert the easting/northing fields
provided by GIAS into equivalent latitude/longitude values, which will
power our location-based search. This is another bit of 'massaging'
which needs to happen before the data can be imported.

I've therefore decided to refactor the GIAS import process so it more
explicitly follows an [ETL][1] (extract, transform, load) pattern:

1. Download the CSV
2. Transform the CSV (filter out unwanted schools)
3. Import the CSV

This reduces the responsibility of the 'import' step by extracting the
filtering logic into a standalone 'transform' step. The importer now
simply needs to import every row of the provided CSV. And it'll make it
easier to change and add to the transformation logic in future.

This refactor doesn't affect the outcome of the import. The high-level
integration test for the entire import still passes as before:

```
bundle exec rspec spec/jobs/gias/sync_all_schools_job_spec.rb
```

One change I have made is in testing for edge cases. Previously the
importer handled edge cases such as "invalid" CSV files where schools
were missing a name or URN. However in practice this never happens: the
GIAS CSV file always has a name for each school. The edge-case handling
was also not comprehensive – it didn't handle cases where the CSV was
missing expected columns or other field values, for example.

Since we haven't seen "invalid" CSV files in practice, it seems like a
low-risk change to remove tests for these edge cases. I'd rather keep
things simple. And then if we end up receiving corrupt CSV files in the
future, we'll be alerted by Sentry with unhandled exceptions when
schools fail to save to the database. This will make the problem more
obvious instead of quietly skipping those schools.

[1]: https://en.wikipedia.org/wiki/Extract,_transform,_load