Fix test failures caused by lat/long variations
The output of the PROJ `cs2cs` command seem to vary _very slightly_
depending on the software version and/or operating system.

This commit tweaks the tests to build in some tolerance for slight
variations in the values. In pratice, these represent differences of
about 2 metres, so they're perfectly acceptable for our use case.