Add auth/failure route
In very edge cases, DSI will send an auth/failure request to our app
with a message.

We need to handle this gracefully so we redirect the user to the sign in
page, sign him our if they are signed in and display a message.

We also send a request to Sentry.