Update Claims::Claim::MentorTrainingForm to show max amount of claimable hours instead of static 20 hours
I've also cleaned up the form object to use the initializer instead of a `updated_mentor_training` variable, which also avoids the need to propagate errors.

The validation has been added at the form level. This means there is a potential case where 2 users are on the check page and have not submitted, they may get away with the submission. To be discussed further for submission prevention.