Add dfe-analytics
This will send data on every web request and whenever a table in
config/analytics is updated/created.

A scheduled job is also going to run every day after midnight to keep in
sync the BigQuery data with our own data.

We will send data to dfe-analytics in Production, Sandbox, Staging and
QA.