Remove placement windows
Placement windows are not something we're intending to use on 'day 1' of
the service.

They will come at a later date, so for now I'm stripping this all out.
There's every likelihood that they'll need to be implemented differently
next time they're introduced – but if not, we can always revert this
change and reinstate placement windows as they were originally.