Implement claim revisions on normal claim form
In testing the claim form we concluded that we cannot edit the same
claim record on the check page when creating a claim.

Issues were found when the user would click the back buttons to go back
to the check page after completing half their changes. The changes they
did were being saved. Even if they backed off from doing them.

To solve this issue and many others, we are implementing a revision
system on the claim.

Creating a revision every time a user wants to edit a claim from the
check page. If they back off half way through their changes we revert
back to the last valid claim revision.

This implementation will increase the audit records and possibly pollute
some of them with `system changes` i.e. changes to `previous_revision_id`
or `next_revision_id` which the user doesn't care about.

We will discuss a solution to this in the future.