Ask for DOB when adding a claim mentor
We want the user to input a date of birth when adding a mentor.
Otherwise its easy to guess your way to a trn and revealing personal
information about teachers.

The teacher record api can do this validation for us we just need to
update the version that we use and pass the DOB to the teacher record
api.

This also updates the Teacher Record api version to accept dob