Remove start_date and end_date from placements
Placement windows, and their associated start dates and end dates, are
not required for 'day 1' of the service.

They will be re-introduced later once we have a better understanding of
exactly how they need to work, and the functionality they'll enable.

Until then I don't think it's useful to keep these hanging around.