Reject claims with invalid mentor training hours
We have a race condition where a user can create a claim for a mentor
before someone else trying to do the same thing.

So you can end up with more mentor training hours per provider than the
maximum 20 hours allowed, per provider.

This PR adds a validation step in the submit and draft controller
actions and redirect the user to a reject page when this happens

This PR adds a validation step in the submit and draft controller
actions and redirect the user to a reject page when this happens.