Disable DfE Analytics for the healthcheck endpoint
DfE Analytics automatically logs every web request received by the
application. This happens because the top-level `ApplicationController`
includes the `DfE::Analytics::Requests` concern, meaning every
controller automatically inherits this behaviour.

For the most part, this is desirable behaviour. As described in the
[dfe-analytics][1] README, the gem provides an _opinionated_ integration
which is designed to send _every_ web request into BigQuery.

However, the `/healthcheck` endpoint is a special case. This endpoint is
hit very frequently by our infrastructure layer to make sure the app is
up, running, and healthy. From checking request logs, it looks like this
is hit once every second.

It's not valuable to record these healthcheck requests in BigQuery. In
fact it's quite wasteful of resources. In just a few days of having this
integration enabled in our QA environment, we've already racked up ~500k
jobs to push these requests into BigQuery (~3600 per hour, assuming no
other traffic).

This change should save us a considerable amount of wasted CPU time,
network traffic and disk usage in the database and BigQuery.

Unfortunately the gem doesn't provide a 'nicer' way to skip request
logging on certain controllers. But since it's all powered by
[controller callbacks][2], we can explicitly skip that callback in the same
way we already do for user authentication.

I thought it was also valuable to add a test to prove that we _do_ send
requests into DfE Analytics for other requests. Otherwise we could
accidentally break the integration and the test suite would continue
passing.

[1]: https://github.com/DFE-Digital/dfe-analytics
[2]: https://github.com/DFE-Digital/dfe-analytics/blob/656807c8ee97e74054ad752970313b3d6f8e86e0/lib/dfe/analytics/requests.rb#L9