Skip ApplicationController on healthcheck endpoint
Refactor the HeartbeatController (which powers the `/healthcheck`
endpoint) so it doesn't inherit from the `ApplicationController`.

This simplifies the controller because it no longer needs to explicitly
skip undesirable functionality such as user authentication or
integration with DfE Analytics.

This controller doesn't need all the magic and niceties of regular
user-facing controllers.

By inheriting from [ActionController::API][1], the controller will now
be as lightweight as possible. Which is ideal for this endpoint because
it gets hit every second or thereabouts.

[1]: https://api.rubyonrails.org/classes/ActionController/API.html