Implement revisions to creating support claims flow
This should fix the issue where a support user edits a draft claim and
with every edit, the draft claim is updated for everyone.

This will only `update` the draft claim when the support user actually
clicks `Update claim`.

They way this works is by using the `internal_draft` status. A generated
revision starts with `internal_draft` as its status and is changed to
draft when the support user finishes editing the draft claim. The
previous revision is then set to `internal_draft`, to not show more than
1 claim per reference.