Fix back link when creating a claim
When creating a claim you can end up on the edit page of a claim, or
mentors list.

These edit pages were programmed to always go to the check page, so when
a user would click them it would try to redirect them to the check page.

This commit fixes this with a reviewed boolean field which is set to
true when the user lands on the check page.

We use this boolean to know where to redirect the user, back to the
check page or through the steps ending on the index page.

Once a user lands on the check page, the back buttons should always
redirect him to the check page.