Add breadcrumbs to our static pages
We want a way for the user to get back to the home page from our static
pages. So the breadcrumbs will contain the current page and a home page.

The home page link will go to the school/claims index page if normal the user
is logged in.

For support users it will redirect to organisations index page. This is
because we redirect from the root path to these pages if a user is
signed in.