Fix support claims spec warning
The claims/support/schools/claims/change_claim_on_check_page spec was
checking against nil values, this is not really good practice.

This PR fixes this by removing the specs doing this as they don't really
add value. The intention was to not create new claims when editing a
claim as a support user but this is done in the support edit_draft_claim_spec