Reload placement for additional subjects after filter
When we filter down the placements we also clip out their additional subjects. The simplest way to resolve this issue is to reload the record when we need the additional subject names. We do not want to reload the record when dealing with the add a placement flow as the record is not saved at this point and no filters can be applied.