Reduce the debounce timeout on autocompletes
In testing, we noticed that the autocomplete suggestions that were
sourced via Ajax felt very slow and clunky. I think there were a few
compounding factors which resulted in this poor UX:

1. The client-side was slow to react to keystrokes and trigger Ajax
   requests for suggestions from the server.
2. There was no visual hint to show users that something was happening.
3. The server was slow to respond to Ajax requests.

This commit addresses Point 1 from the list.

I've reduced the debounce timeout on Ajax-sourced autocompletes from
900ms to 200ms.

This should make the autocomplete feel more responsive by asking the
server for suggestions sooner after the user has typed into the box.

There is always a risk that reducing the debouncing timeout will cause
excessive load on the server-side. However we have already made changes
to improve the server response time in #735.

Additionally, the `request` helper method that we use for Ajax requests
will cancel existing in-flight requests if a new one is triggered. This
should reduce our impact on the server by making sure we cancel stale
requests, which hopefully frees up resources more quickly on the
server-side.