Optimise the size of autocomplete responses
I've tweaked the autocomplete suggestion controllers to reduce the
amount of data which is transferred over the wire. This should result in
autocomplete suggestions appearing more quickly in the UI.

Changes I've made:

1. Only return fields that are needed

   Originally, every attribute on the School or Provider model was being
   returned in the Ajax response. However the UI only required two or
   three attributes:

   Providers:
    - id
    - name
    - code

   Schools:
    - id
    - name
    - town
    - postcode

   So I've limited the responses to just these fields.

2. Limit the number of suggestions returned

   There was no limit to the number of suggestions returned. Given the
   number of schools in the GIAS dataset, it'd be quite easy to generate
   a request which would return hundreds or thousands of schools in one
   response.

   In practice, and knowing how these results will be presented, it's
   unlikely users would ever need to see more than 50 suggestions at a
   time.

This has a huge impact on the response size of Ajax calls, and in turn
should boost performance by reducing load on both the server and client.

For example, I tried typing "Al" into the autocomplete.

Before this change, the server responded with 670 suggestions.
The response size was 871 KB.

Following this change, the server responds with 50 suggestions.
The response size is 8 KB.

That's a 100x reduction in response size, which should translate nicely
into real-world improvements to performance and responsiveness.