De-flake test
I've hardcoded the school and provider names so they're always
guaranteed to be unique. Without this, the tests would occasionally
fail because 'school' and 'another_school', or 'provider' and
'another_provider', could end up sharing the same name.

This would cause the test to fail because we assert on 'another_school's
name _not being present_ on the page.