Add test coverage for user deletion
I've added request specs to prove that users must be authorized when
deleting other users.

It also proves that users cannot delete themselves.

I've added this because I noticed we weren't covering the "support user
cannot delete themselves" scenario, and so it was possible to remove the
authorization code and the tests would still pass.