Remove Placements::Support::ApplicationController
This commit removes the parent controller
Placements::Support::ApplicationController which controllers in the
support console previously needed to extend.

Since we handle support user authorisation in the new
Placements::ApplicationController now, we no longer need a dedicated
ApplicationController for the support console.

It can always be added again later if needed, but for now it doesn't
serve any purpose. So it's one layer of abstraction that we can safely
remove. 🎉