Import proposed to be closed schools from GIAS
There are a small amout of schools that are proposed to be closed in the
future but are currently still open. These schools should still be able
to claim funding for their mentors.

So this commit changes the GIAS importer to accept schools that are
proposed to be closed.

Work will follow to restrict these schools after they have actually been
closed.