Refactor how the organisation switcher works
I've refactored how the organisation switcher works to make sure it
always sets `session[:placements_provider_id]` when switching between
organisations.

This was the source of a bug whereby users could switch from one
provider to another, but then clicking on the 'Placements' tab from the
new provider would still load the previous one.

It also solves an unexpected behaviour where the
`placements_organisation_path` helper would update the user's session
when called. This was an unexpected side effect which caused the session
to update while listing out links on the organisations page.