Block PII in school contacts from DfE Analytics
These are known as 'placements contacts' in the app, and they contain
names and email addresses of people for providers to contact.

Even though they're technically PII so could go into the
analytics_pii.yml config, we've decided to put it into
analytics_blocklist.yml instead, so it's never sent to DfE Analytics in
the first place.