Enable status persistence for RSpec examples
This allows developers to use the `--only-failures` and `--next-failure`
flags when running RSpec. These make it much quicker when you want a
quicker feedback loop while fixing disparate tests that are failing
across the suite.

For example, it's handy when a content change ends up breaking various
unit and system specs. 🥲