Store user sessions in the database
By default, Rails will store user sessions in an encrypted cookie. This
works fine until you need to begin storing larger amounts of data in the
session, because cookies have a ~4 KB size limit.

Instead, I've set up the Active Record session store. This will persist
user sessions in the database, where we don't have the same size
constraints.

This is important because we use sessions to remember the state of
multi-page wizard forms, such as the "Add a placement" form. Submitted
form data is persisted in the session until the user hits the "Save"
button on the final "Check your answers" step. That way we don't fill up
the "placements" database table with partially-completed records.