Trim old sessions from the database
Now that we store user sessions in the database, it's important we have
a maintenance task that will regularly clean up old sessions.  Without
this, the sessions table would fill up with loads of old sessions long
past their useful life.

Thankfully it's easy for us to schedule recurring jobs using Good Job.
It has meant I've needed to re-implement the tidy-up Rake task as an
Active Job instead. But I'm expecting this to be one of those things
that we 'set and forget'. It likely won't need to change often, if
at all.

As per the DfE Sign-in implementation, user sessions expire after two
hours of inactivity. Therefore I've configured the job to prune any
sessions older than two hours.