Always use strings for session keys
Sessions in Rails are a bit weird. Annoyingly, Rails session stores tend
to leak underlying implementation details because they differ in the way
they handle hash keys which are symbols vs. those which are strings.

Ideally the Rails session would be a hash with indifferent access.
Indeed, I think it actually _is_ when using the cookie session store.
But that's not the case for the Active Record session store.

I've therefore gone through the codebase and made sure we're
consistently using strings for session keys. This is the safest approach
when reading and writing to sessions, because string keys tend to be
treated the same regardless of the storage adapter used.